* much more warning fixes

+++ src/ta3d/src/ingame/menus/intro.cpp
@@ -27,9 +27,9 @@
 
 
 //!
-# define TA3D_INTRO_TOP (630.0f * SCREEN_H / 1024.0f)
+# define TA3D_INTRO_TOP (630.0f * float(SCREEN_H) / 1024.0f)
 
-# define TA3D_INTRO_BOTTOM (950.0f * SCREEN_H / 1024.0f)
+# define TA3D_INTRO_BOTTOM (950.0f * float(SCREEN_H) / 1024.0f)
 
 # define TA3D_INTRO_MAX_LINES ((unsigned int)((TA3D_INTRO_BOTTOM - TA3D_INTRO_TOP) / pCurrentFontHeight) + 2)
 
@@ -95,7 +95,7 @@ namespace Menus
 
 	bool Intro::maySwitchToAnotherMenu()
 	{
-		pDelta += TA3D_INTRO_SPEED * (msec_timer - pScrollTimer) * 0.001f;
+		pDelta += TA3D_INTRO_SPEED * float(msec_timer - pScrollTimer) * 0.001f;
 		pScrollTimer = msec_timer;
 		if (pDelta > pCurrentFontHeight)
 		{
@@ -118,7 +118,7 @@ namespace Menus
 
 		// Load all text files
 		TA3D_FILE::Load(pContent, "intro" + Paths::SeparatorAsString + I18N::Translate("en.ta3d.txt"), 5 * 1024 /* Max 5Ko */, false);
-		pContentSize = pContent.size();
+		pContentSize = unsigned(pContent.size());
 		if (pContentSize == TA3D_INTRO_MAX_LINES)
 		{
 			pContent.push_back("Welcome to TA3D !");
@@ -153,13 +153,13 @@ namespace Menus
 		// Clear the screen
 		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 		// Background
-		gfx->drawtexture(pBackgroundTexture, 0.0f, 0.0f, SCREEN_W, SCREEN_H);
+		gfx->drawtexture(pBackgroundTexture, 0.0f, 0.0f, float(SCREEN_W), float(SCREEN_H));
 
 		glEnable(GL_BLEND);
 		glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
-		float fw = SCREEN_W / 1280.0f;
-		//float fh = SCREEN_H / 1024.0f;
+		float fw = float(SCREEN_W) / 1280.0f;
+		//float fh = float(SCREEN_H) / 1024.0f;
 
 		// The text itself
 		glColor4ub(0xFF, 0xFF, 0xFF, 0xFF);
@@ -167,7 +167,7 @@ namespace Menus
 		{
 			unsigned int indx = 0;
 			for (unsigned int i = pStartIndex; i < pContentSize && indx < TA3D_INTRO_MAX_LINES; ++i, ++indx)
-				gfx->print(Gui::gui_font, 220.0f * fw, TA3D_INTRO_TOP + (indx-1) * pCurrentFontHeight - pDelta, 0.0f, pContent[i]);
+				gfx->print(Gui::gui_font, 220.0f * fw, TA3D_INTRO_TOP + float(indx - 1) * pCurrentFontHeight - pDelta, 0.0f, pContent[i]);
 		}
 
 		glBlendFunc(GL_ONE_MINUS_SRC_ALPHA,GL_SRC_ALPHA);
@@ -176,27 +176,27 @@ namespace Menus
 
 		glColor4ub(0xFF, 0xFF, 0xFF, 0);
 
-		glTexCoord2f(0.0f, (TA3D_INTRO_TOP - 60.0f) / SCREEN_H);                     glVertex2f(0.0f, TA3D_INTRO_TOP - 60.0f);
-		glTexCoord2f(1.0f, (TA3D_INTRO_TOP - 60.0f) / SCREEN_H);                     glVertex2f(SCREEN_W, TA3D_INTRO_TOP - 60.0f);
-		glTexCoord2f(1.0f, TA3D_INTRO_TOP / SCREEN_H);                               glVertex2f(SCREEN_W, TA3D_INTRO_TOP);
-		glTexCoord2f(0.0f, TA3D_INTRO_TOP / SCREEN_H);                               glVertex2f(0.0f, TA3D_INTRO_TOP);
+		glTexCoord2f(0.0f, (TA3D_INTRO_TOP - 60.0f) / float(SCREEN_H));                     glVertex2f(0.0f, TA3D_INTRO_TOP - 60.0f);
+		glTexCoord2f(1.0f, (TA3D_INTRO_TOP - 60.0f) / float(SCREEN_H));                     glVertex2f(float(SCREEN_W), TA3D_INTRO_TOP - 60.0f);
+		glTexCoord2f(1.0f, TA3D_INTRO_TOP / float(SCREEN_H));                               glVertex2f(float(SCREEN_W), TA3D_INTRO_TOP);
+		glTexCoord2f(0.0f, TA3D_INTRO_TOP / float(SCREEN_H));                               glVertex2f(0.0f, TA3D_INTRO_TOP);
 
-		glTexCoord2f(0.0f, TA3D_INTRO_TOP / SCREEN_H);                               glVertex2f(0.0f, TA3D_INTRO_TOP);
-		glTexCoord2f(1.0f, TA3D_INTRO_TOP / SCREEN_H);                               glVertex2f(SCREEN_W, TA3D_INTRO_TOP);
+		glTexCoord2f(0.0f, TA3D_INTRO_TOP / float(SCREEN_H));                               glVertex2f(0.0f, TA3D_INTRO_TOP);
+		glTexCoord2f(1.0f, TA3D_INTRO_TOP / float(SCREEN_H));                               glVertex2f(float(SCREEN_W), TA3D_INTRO_TOP);
 		glColor4ub(0xFF,0xFF,0xFF,0xFF);
-		glTexCoord2f(1.0f, (TA3D_INTRO_TOP + 2 * pCurrentFontHeight) / SCREEN_H);    glVertex2f(SCREEN_W, TA3D_INTRO_TOP + 2 * pCurrentFontHeight);
-		glTexCoord2f(0.0f, (TA3D_INTRO_TOP + 2 * pCurrentFontHeight) / SCREEN_H);    glVertex2f(0.0f, TA3D_INTRO_TOP + 2 * pCurrentFontHeight);
+		glTexCoord2f(1.0f, (TA3D_INTRO_TOP + 2 * pCurrentFontHeight) / float(SCREEN_H));    glVertex2f(float(SCREEN_W), TA3D_INTRO_TOP + 2 * pCurrentFontHeight);
+		glTexCoord2f(0.0f, (TA3D_INTRO_TOP + 2 * pCurrentFontHeight) / float(SCREEN_H));    glVertex2f(0.0f, TA3D_INTRO_TOP + 2 * pCurrentFontHeight);
 
-		glTexCoord2f(0.0f, (TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight) / SCREEN_H); glVertex2f(0.0f, TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight);
-		glTexCoord2f(1.0f, (TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight) / SCREEN_H); glVertex2f(SCREEN_W, TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight);
+		glTexCoord2f(0.0f, (TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight) / float(SCREEN_H)); glVertex2f(0.0f, TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight);
+		glTexCoord2f(1.0f, (TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight) / float(SCREEN_H)); glVertex2f(float(SCREEN_W), TA3D_INTRO_BOTTOM - 2 * pCurrentFontHeight);
 		glColor4ub(0xFF, 0xFF, 0xFF, 0);
-		glTexCoord2f(1.0f, TA3D_INTRO_BOTTOM / SCREEN_H);                            glVertex2f(SCREEN_W, TA3D_INTRO_BOTTOM);
-		glTexCoord2f(0.0f, TA3D_INTRO_BOTTOM / SCREEN_H);                            glVertex2f(0.0f, TA3D_INTRO_BOTTOM);
+		glTexCoord2f(1.0f, TA3D_INTRO_BOTTOM / float(SCREEN_H));                            glVertex2f(float(SCREEN_W), TA3D_INTRO_BOTTOM);
+		glTexCoord2f(0.0f, TA3D_INTRO_BOTTOM / float(SCREEN_H));                            glVertex2f(0.0f, TA3D_INTRO_BOTTOM);
 
-		glTexCoord2f(0.0f, TA3D_INTRO_BOTTOM / SCREEN_H);                            glVertex2f(0.0f, TA3D_INTRO_BOTTOM);
-		glTexCoord2f(1.0f, TA3D_INTRO_BOTTOM / SCREEN_H);                            glVertex2f(SCREEN_W, TA3D_INTRO_BOTTOM);
-		glTexCoord2f(1.0f, (TA3D_INTRO_BOTTOM + 60.0f) / SCREEN_H);                  glVertex2f(SCREEN_W, TA3D_INTRO_BOTTOM + 60.0f);
-		glTexCoord2f(0.0f, (TA3D_INTRO_BOTTOM + 60.0f) / SCREEN_H);                  glVertex2f(0.0f, TA3D_INTRO_BOTTOM + 60.0f);
+		glTexCoord2f(0.0f, TA3D_INTRO_BOTTOM / float(SCREEN_H));                            glVertex2f(0.0f, TA3D_INTRO_BOTTOM);
+		glTexCoord2f(1.0f, TA3D_INTRO_BOTTOM / float(SCREEN_H));                            glVertex2f(float(SCREEN_W), TA3D_INTRO_BOTTOM);
+		glTexCoord2f(1.0f, (TA3D_INTRO_BOTTOM + 60.0f) / float(SCREEN_H));                  glVertex2f(float(SCREEN_W), TA3D_INTRO_BOTTOM + 60.0f);
+		glTexCoord2f(0.0f, (TA3D_INTRO_BOTTOM + 60.0f) / float(SCREEN_H));                  glVertex2f(0.0f, TA3D_INTRO_BOTTOM + 60.0f);
 
 		glEnd();
 		glDisable(GL_BLEND);
